Rename file for MinimumWageFlow match its class name
