+++ lib/smart_answer_flows/shared/minimum_wage_flow.rb
